Arrows kill Wumpus from any dir
Test shooting Wumpus from 4 directions instead of 1.
This was a response to the concern at 1:11:04 in
Clean Code Episode 37 - Elaborating the Requirements.

I don't know if this add much value, and arguably could
make the tests more difficult to read as it becomes
more and more like a program. If there was a need for a
rigid test, perhaps the test details should be placed far
from non-programmers view, to make it read like english.
Perhaps even a Unit test?