Implement missing test
I discovered an implemented feature that was not tested.
Then, I had to choose between two paths:
    1. Deleting the feature (if it's not in the requirements, then it
       should not be in the code), or
    2. Adding a test, to justify the existence of the feature.

I chose to add a new test (2), because I wanted to practice with FitNesse,
but in a real project I would consider the other option a little more (an
arrow can travel at most 103 caverns? and it isn't in the requirements?
sounds suspicious...)